+++ 数据表/t_order_item.sql
@@ -6,8 +6,8 @@ create table t_order_item
 	product_id int not null comment '商品id',
 	order_id int not null comment '订单id',
 	num int not null comment '商品购买数量',
-	price bigint not null comment '商品单价(单位：分)',
-	total_fee bigint not null comment '商品总金额',
+	price decimal not null comment '商品单价(单位：分)',
+	total_fee decimal not null comment '商品总金额',
 	constraint t_order_item_t_product_info_id_fk
 		foreign key (product_id) references t_product_info (id),
 	constraint t_order_item_t_order_id_fk
@@ -23,4 +23,6 @@ create index t_order_item_t_order_id_fk
 create index t_order_item_t_product_info_id_fk
 	on t_order_item (product_id)
 ;
+
+
